Move random to contract_comm
